chore: move export policy validation
Use an internal package to handle the validation of the export policy.
